requirements.txt
# MT5 Connection (native — requires Windows + MT5 terminal running)
MetaTrader5>=5.0.45

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0

# Data & Indicators
pandas>=2.1.0
numpy>=1.26.0
ta>=0.11.0           # Technical analysis indicators

# AI
groq>=0.4.0          # Groq API (free, fast)
google-generativeai>=0.8.0  # Gemini API (free tier)
anthropic>=0.40.0    # Claude API
openai>=1.6.0        # OpenAI API (fallback)

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
